apps/api-flask/pyproject.toml
[project]
name = "api-flask"
version = "0.1.0"
description = "Python flask based todo application back-end"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "flask-cors>=5.0.0",
    "flask-jwt-extended>=4.7.1",
    "flask-marshmallow>=1.2.1",
    "marshmallow-sqlalchemy>=1.1.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
    "argon2-cffi>=23.1.0",
    "email-validator>=2.2.0",
    "flask-swagger-ui>=4.11.1",
    "pyyaml>=6.0.2",
    "click>=8.1.8",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[project.scripts]
flask-api = "main:app"
manage = "scripts.manage:cli"

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
